+++ itemsystem/src/main/java/com/example/itemsystem/dto/OrderHistoryApiDto.java
@@ -1,21 +1,14 @@
+package com.example.itemsystem.dto;
 
-	package com.example.itemsystem.entity;
+import lombok.Data;
 
-	import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-	@Entity
-	public class OrderHistoryApiEntity {
-		  @Id
-		    @GeneratedValue(strategy = GenerationType.IDENTITY)
-		    private Long id;
-
-		    private String productNname;
-		    private String managerName;
-		    private int orderQuantity;
-		    private int totalAmount;
-
-	}
+@Data
+public class OrderHistoryApiDto {
+	  
+	private Long id;
+	private String productNname;
+	private String managerName;
+	private Long orderQuantity;
+	private Long totalAmount;
+}
