+++ itemsystem/src/main/java/com/example/itemsystem/entity/AdminEntity.java
@@ -1,8 +1,9 @@
 package com.example.itemsystem.entity;
 
 import java.time.LocalDateTime;
-import java.util.Optional;
+import java.util.List;
 
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -10,6 +11,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
@@ -19,81 +21,70 @@ import lombok.Data;
 @Data
 @Table(name = "admin")
 public class AdminEntity {
-	    @Id
-	    @GeneratedValue(strategy=GenerationType.IDENTITY)
-	    @Column(name = "id")
-	    private Long id;
-	   
-	    @ManyToOne
-	    @JoinColumn(name = "shop_id", referencedColumnName = "id")
-	    private ShopEntity shop;
-	   
-	    @Column(name = "last_name", nullable = false)
-	    private String lastName;
-
-	    @Column(name = "first_name", nullable = false)
-	    private String firstName;
-
-	    @Column(name = "email", nullable = false)
-	    private String email;
-	    
-	    @ManyToOne
-	    @JoinColumn(name = "role_id", referencedColumnName = "id")
-	    private RoleEntity role;
-
-	    @ManyToOne
-	    @JoinColumn(name = "permission_id", referencedColumnName = "id")
-	    private PermissionEntity permission;
-
-	    @Column(name = "phone_number", nullable = false)
-	    private String phoneNumber;
-
-	    @Column(name = "password", nullable = false)
-	    private String password;
-	    
-	    public Long getShopId() {
-	        return (shop != null) ? shop.getId() : null;
+	@Id
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	@Column(name = "id")
+    private Long id;
+   
+    @ManyToOne
+    @JoinColumn(name = "shop_id", referencedColumnName = "id")
+    private ShopEntity shop;
+   
+    @Column(name = "last_name", nullable = false)
+	private String lastName;
+	
+	@Column(name = "first_name", nullable = false)
+	private String firstName;
+	
+	@Column(name = "email", nullable = false)
+	private String email;
+	
+	@ManyToOne
+	@JoinColumn(name = "role_id", referencedColumnName = "id")
+	private RoleEntity role;
+	
+	@ManyToOne
+	@JoinColumn(name = "permission_id", referencedColumnName = "id")
+	private PermissionEntity permission;
+	
+	@Column(name = "phone_number", nullable = false)
+	private String phoneNumber;
+	
+	@Column(name = "password", nullable = false)
+	private String password;
+	
+	public Long getShopId() {
+	    return (shop != null) ? shop.getId() : null;
+	}
+	
+	public Long getRoleId() {
+	    return (role != null) ? role.getId() : null;
+	}
+	public Long getPermissionId() {
+	    return (permission != null) ? permission.getId() : null;
+	}
+	
+	@Column(name = "created_at", nullable =false ,updatable = false)
+	private LocalDateTime created_at;
+	
+	@Column(name = "updated_at", nullable = false)
+	private LocalDateTime updated_at;
+	
+	@PrePersist
+	protected void onCreate() {
+	    if (this.created_at == null) {
+	        this.created_at = LocalDateTime.now();
 	    }
-
-	    public Long getRoleId() {
-	        return (role != null) ? role.getId() : null;
-	    }
-	    public Long getPermissionId() {
-	        return (permission != null) ? permission.getId() : null;
-	    }
-	    
-	    @Column(name = "created_at", nullable =false ,updatable = false)
-	    private LocalDateTime created_at;
-	    
-	    @Column(name = "updated_at", nullable = false)
-	    private LocalDateTime updated_at;
-	    
-	    @PrePersist
-	    protected void onCreate() {
-	        if (this.created_at == null) {
-	            this.created_at = LocalDateTime.now();
-	        }
-	        if (this.updated_at == null) {
-	            this.updated_at = LocalDateTime.now();
-	        }
-	    }
-
-	    @PreUpdate
-	    protected void onUpdate() {
+	    if (this.updated_at == null) {
 	        this.updated_at = LocalDateTime.now();
 	    }
-
-		public Optional<ShopEntity> map(Object object) {
-			// TODO 自動生成されたメソッド・スタブ
-			return null;
-		}
-
-		public AdminEntity orElseThrow(Object object) {
-			// TODO 自動生成されたメソッド・スタブ
-			return null;
-		}
-
+	}
 	
-	    
-
+	@PreUpdate
+	protected void onUpdate() {
+	    this.updated_at = LocalDateTime.now();
+	}
+	
+	@OneToMany(mappedBy = "admin", cascade = CascadeType.ALL)
+	   private List<OrderHistoryEntity> orderHistory;
 }
