+++ itemsystem/src/main/java/com/example/itemsystem/entity/ItemSmallCategoryEntity.java
@@ -19,36 +19,40 @@ import lombok.Data;
 @Table(name = "item_small_category")
 public class ItemSmallCategoryEntity {
 	@Id
-    @GeneratedValue(strategy=GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	@Column(name = "id")
+	private Long id;
 	
 	@ManyToOne
 	@JoinColumn(name = "item_under_category_id", referencedColumnName = "id")
-    private ItemUnderCategoryEntity itemUnderCategory;
+	private ItemUnderCategoryEntity itemUnderCategory;
 	
 	@Column(name = "name")
 	private String name;
 	
-    @Column(name = "created_at", nullable =false ,updatable = false)
-    private LocalDateTime created_at;
-    
-    @Column(name = "updated_at", nullable = false)
-    private LocalDateTime updated_at;
-    
-    @PrePersist
-    protected void onCreate() {
-        if (this.created_at == null) {
-            this.created_at = LocalDateTime.now();
-        }
-        if (this.updated_at == null) {
-            this.updated_at = LocalDateTime.now();
-        }
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-        this.updated_at = LocalDateTime.now();
-    }
-
+	@Column(name = "created_at", nullable =false ,updatable = false)
+	private LocalDateTime created_at;
+	
+	@Column(name = "updated_at", nullable = false)
+	private LocalDateTime updated_at;
+	
+	@PrePersist
+	protected void onCreate() {
+	    if (this.created_at == null) {
+	        this.created_at = LocalDateTime.now();
+	    }
+	    if (this.updated_at == null) {
+	        this.updated_at = LocalDateTime.now();
+	    }
+	}
+	
+	@PreUpdate
+	protected void onUpdate() {
+	    this.updated_at = LocalDateTime.now();
+	}
+	
+	public String getSmallCategoryName() {
+		// TODO 自動生成されたメソッド・スタブ
+		return name;
+	}
 }
