+++ itemsystem/src/main/java/com/example/itemsystem/entity/ManufacturesEntity.java
@@ -17,34 +17,31 @@ import lombok.Data;
 @Data
 @Table(name = "manufactures")
 public class ManufacturesEntity {
-	
 	@Id
-    @GeneratedValue(strategy=GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	@Column(name = "id")
+	private Long id;
 	
 	@Column(name = "name")
 	private String name;
 	
-    @Column(name = "created_at", nullable =false ,updatable = false)
-    private LocalDateTime created_at;
-    
-    @Column(name = "updated_at", nullable = false)
-    private LocalDateTime updated_at;
-    
-    @PrePersist
-    protected void onCreate() {
-        if (this.created_at == null) {
-            this.created_at = LocalDateTime.now();
-        }
-        if (this.updated_at == null) {
-            this.updated_at = LocalDateTime.now();
-        }
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-        this.updated_at = LocalDateTime.now();
-    }
-
+	@Column(name = "created_at", nullable =false ,updatable = false)
+	private LocalDateTime created_at;
+	
+	@Column(name = "updated_at", nullable = false)
+	private LocalDateTime updated_at;
+	
+	@PrePersist
+	protected void onCreate() {
+	    if (this.created_at == null) {
+	        this.created_at = LocalDateTime.now();
+	    }
+	    if (this.updated_at == null) {
+	        this.updated_at = LocalDateTime.now();
+	    }
+	}
+	@PreUpdate
+	protected void onUpdate() {
+	    this.updated_at = LocalDateTime.now();
+	}
 }
