+++ itemsystem/src/main/java/com/example/itemsystem/entity/OrderHistoryEntity.java
@@ -19,49 +19,73 @@ import lombok.Data;
 @Table(name = "order_history")
 
 public class OrderHistoryEntity {
-	 @Id
-	    @GeneratedValue(strategy=GenerationType.IDENTITY)
-	    @Column(name = "id")
-	    private Long id;
-	   
-	 	@ManyToOne
-	    @JoinColumn(name = "admin_id", referencedColumnName = "id")
-	    private AdminEntity admin;
-	 
-	    @ManyToOne
-	    @JoinColumn(name = "shop_id", referencedColumnName = "id")
-	    private ShopEntity shop;
-	    
-	    @ManyToOne
-	    @JoinColumn(name = "item_id", referencedColumnName = "id")
-	    private ItemEntity item;
-	    
-	    @Column(name = "number_of_orders")
-	    private Long numberOfOrders;
-	    
-	    @Column(name = "total_amount")
-	    private Long totalAmount;
-
-	    
-	    @Column(name = "created_at", nullable =false ,updatable = false)
-	    private LocalDateTime created_at;
-	    
-	    @Column(name = "updated_at", nullable = false)
-	    private LocalDateTime updated_at;
-	    
-	    @PrePersist
-	    protected void onCreate() {
-	        if (this.created_at == null) {
-	            this.created_at = LocalDateTime.now();
-	        }
-	        if (this.updated_at == null) {
-	            this.updated_at = LocalDateTime.now();
-	        }
+	@Id
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	@Column(name = "id")
+    private Long id;
+   
+ 	@ManyToOne
+    @JoinColumn(name = "admin_id", referencedColumnName = "id")
+    private AdminEntity admin;
+ 
+    @ManyToOne
+    @JoinColumn(name = "shop_id", referencedColumnName = "id")
+	private ShopEntity shop;
+	
+	@ManyToOne
+	@JoinColumn(name = "item_id", referencedColumnName = "id")
+	private ItemEntity item;
+	
+	@Column(name = "number_of_orders")
+	private Long numberOfOrders;
+	
+	@Column(name = "total_amount")
+	private Long totalAmount;
+	
+	
+	@Column(name = "created_at", nullable =false ,updatable = false)
+	private LocalDateTime created_at;
+	
+	@Column(name = "updated_at", nullable = false)
+	private LocalDateTime updated_at;
+	
+	@PrePersist
+	protected void onCreate() {
+	    if (this.created_at == null) {
+	        this.created_at = LocalDateTime.now();
 	    }
-
-	    @PreUpdate
-	    protected void onUpdate() {
+	    if (this.updated_at == null) {
 	        this.updated_at = LocalDateTime.now();
 	    }
-
+	}
+	
+	@PreUpdate
+	protected void onUpdate() {
+	    this.updated_at = LocalDateTime.now();
+	}
+	
+	public String getProductName() {
+		// TODO 自動生成されたメソッド・スタブ
+		return null;
+	}
+	
+	public String getManagerName() {
+		// TODO 自動生成されたメソッド・スタブ
+		return null;
+	}
+	
+	public Long getOrderQuantity() {
+		// TODO 自動生成されたメソッド・スタブ
+		return numberOfOrders;
+	}
+	
+	public Long getTotalAmount(Long totalAmout) {
+		// TODO 自動生成されたメソッド・スタブ
+		return totalAmout;
+	}
+	
+	public Long getId(Long id) {
+		// TODO 自動生成されたメソッド・スタブ
+		return id;
+	}	
 }
