+++ itemsystem/src/main/java/com/example/itemsystem/form/ItemForm.java
@@ -8,30 +8,30 @@ import lombok.Data;
 @Data
 public class ItemForm {
 
-private Long id;
-
-@Min(1)
-@NotNull
-private Long manufacturesId;
-
-@Min(1)
-@NotNull
-private Long itemSmallCategoryId;
-
-@NotBlank
-private String name;
-
-@NotBlank
-private String productDescription;
-
-@NotBlank
-private String image;
-
-@Min(1)
-@NotNull
-private Long purchasePrice;
-
-@Min(1)
-@NotNull
-private Long retailPrice;
+	private Long id;
+	
+	@Min(1)
+	@NotNull
+	private Long manufacturesId;
+	
+	@Min(1)
+	@NotNull
+	private Long itemSmallCategoryId;
+	
+	@NotBlank
+	private String name;
+	
+	@NotBlank
+	private String productDescription;
+	
+	@NotBlank
+	private String image;
+	
+	@Min(1)
+	@NotNull
+	private Long purchasePrice;
+	
+	@Min(1)
+	@NotNull
+	private Long retailPrice;
 }
