+++ itemsystem/src/main/java/com/example/itemsystem/repository/CategoryApiRepository.java
@@ -2,8 +2,8 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entity.CategoryApiEntity;
+import com.example.itemsystem.entity.ItemEntity;
 
-public interface CategoryApiRepository extends JpaRepository<CategoryApiEntity, Long>{
+public interface CategoryApiRepository extends JpaRepository<ItemEntity,Long>{
 
 }
