+++ itemsystem/src/main/java/com/example/itemsystem/repository/ItemApiRepository.java
@@ -2,8 +2,8 @@ package com.example.itemsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.itemsystem.entity.ItemApiEntity;
+import com.example.itemsystem.entity.ShopItemEntity;
 
-public interface ItemApiRepository  extends JpaRepository<ItemApiEntity, Long>{
+public interface ItemApiRepository  extends JpaRepository<ShopItemEntity, Long>{
 
 }
