+++ itemsystem/src/main/java/com/example/itemsystem/repository/ShopItemRepository.java
@@ -1,5 +1,6 @@
 package com.example.itemsystem.repository;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,7 +13,7 @@ public interface ShopItemRepository extends JpaRepository<ShopItemEntity, Long>{
 	Optional<ShopItemEntity> findByShopAndItem(ShopEntity shop, ItemEntity item);
 
 	ShopItemEntity  findByShop_IdAndItem_Id(Long shopId, Long id);
+
+	List<ShopItemEntity> findByShop_Id(Long shopId);
 	
-
-
 }
