+++ itemsystem/src/main/java/com/example/itemsystem/service/ApiService.java
@@ -1,0 +1,15 @@
+package com.example.itemsystem.service;
+
+import java.util.List;
+
+import com.example.itemsystem.dto.CategoryApiDto;
+import com.example.itemsystem.dto.ItemApiDto;
+
+public interface ApiService {
+
+	List<CategoryApiDto> getCategories();
+
+	List<ItemApiDto> getItems();
+
+
+}
