--- itemsystem/src/main/java/com/example/itemsystem/service/CategoryApiService.java
@@ -1,5 +1,0 @@
-package com.example.itemsystem.service;
-
-public interface CategoryApiService {
-
-}
