--- itemsystem/src/main/java/com/example/itemsystem/service/ItemApiService.java
@@ -1,5 +1,0 @@
-package com.example.itemsystem.service;
-
-public interface ItemApiService {
-
-}
