+++ itemsystem/src/main/java/com/example/itemsystem/service/OlderHistoryService.java
@@ -1,5 +1,11 @@
 package com.example.itemsystem.service;
 
+import java.util.List;
+
+import com.example.itemsystem.entity.OrderHistoryEntity;
+
 public interface OlderHistoryService {
+
+	List<OrderHistoryEntity> getAllItem();
 
 }
