+++ itemsystem/src/main/resources/templates/orderhistory/list.html
@@ -3,11 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>管理者一覧</title>
+    <title>商品発注履歴</title>
     <link rel="stylesheet" href="/css/style.css">
 </head>
 <body>
-    <h1>管理者一覧</h1>
+    <h1>商品発注履歴</h1>
 
     <table>
         <thead>
@@ -16,7 +16,7 @@
                 <th>発注者</th>
                 <th>商品名</th>
 				<th>発注数</th>
-11              <th>合計金額</th>
+                <th>合計金額</th>
                 <th>作成日時</th>
                 <th>更新日時</th>
