+++ itemsystem/src/main/resources/templates/shop/detail.html
@@ -50,7 +50,7 @@
         </form>
 
         <form action="/shop/list" method="post" th:action="@{/shop/list}">
-            <button type="submit">店舗画面へ戻る</button>
+            <button type="submit">店舗一覧画面へ戻る</button>
         </form>
     </div>
 </body>
